pyproject.toml
[tool.poetry]
name = "arxiv-to-podcast"
version = "0.1.0"
description = "Transform academic research papers into engaging podcast conversations"
authors = ["Kaushik Bokka <kaushikbokka@gmail.com>"]
readme = "README.md"
packages = [
    {include = "arxiv_to_podcast", from="."}
]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
pydantic = "^2.7.0"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"
PyPDF2 = "3.0.1"
langgraph = "0.0.15"
chromadb = "0.4.22"
azure-cognitiveservices-speech = "1.34.0"
pydub = "0.25.1"
langchain_openai = "*"
pymupdf = "*"
langchain_community = "*"
langchain_text_splitters = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
